broke argument scanner out into its own file
