+++ src/server/frontend_wayland/generator/wrapper_generator.cpp
@@ -18,7 +18,12 @@
 
 #include "wrapper_generator.h"
 #include "emitter.h"
+#include "argument.h"
 
+#include <libxml++/libxml++.h>
+#include <functional>
+#include <vector>
+#include <experimental/optional>
 #include <iostream>
 #include <unordered_map>
 #include <unordered_set>
@@ -131,56 +136,6 @@ std::string camel_case_string(std::string const& name)
     return camel_cased_name;
 }
 
-struct ArgumentTypeDescriptor
-{
-    std::string cpp_type;
-    std::string c_type;
-    std::experimental::optional<std::vector<std::string>> converter;
-};
-
-std::vector<std::string> fd_converter{
-    "mir::Fd $NAME_resolved{$NAME};"
-};
-
-std::vector<std::string> optional_object_converter{
-    "std::experimental::optional<struct wl_resource*> $NAME_resolved;",
-    "if ($NAME != nullptr)",
-    "{",
-    "    $NAME_resolved = $NAME;",
-    "}"
-};
-
-std::vector<std::string> optional_string_converter{
-    "std::experimental::optional<std::string> $NAME_resolved;",
-    "if ($NAME != nullptr)",
-    "{",
-    "    $NAME_resolved = std::experimental::make_optional<std::string>($NAME);",
-    "}"
-};
-
-std::unordered_map<std::string, ArgumentTypeDescriptor const> type_map = {
-    { "uint", { "uint32_t", "uint32_t", {} }},
-    { "int", { "int32_t", "int32_t", {} }},
-    { "fd", { "mir::Fd", "int", { fd_converter }}},
-    { "object", { "struct wl_resource*", "struct wl_resource*", {} }},
-    { "string", { "std::string const&", "char const*", {} }},
-    { "new_id", { "uint32_t", "uint32_t", {} }}
-};
-
-std::unordered_map<std::string, ArgumentTypeDescriptor const> optional_type_map = {
-    { "object", { "std::experimental::optional<struct wl_resource*> const&", "struct wl_resource*", { optional_object_converter }}},
-    { "string", { "std::experimental::optional<std::string> const&", "char const*", { optional_string_converter} }},
-};
-
-bool parse_optional(xmlpp::Element const& arg)
-{
-    if (auto allow_null = arg.get_attribute("allow-null"))
-    {
-        return allow_null->get_value() == "true";
-    }
-    return false;
-}
-
 void emit_indented_lines(std::ostream& out, std::string const& indent,
                          std::initializer_list<std::initializer_list<std::string>> lines)
 {
@@ -196,54 +151,6 @@ void emit_indented_lines(std::ostream& out, std::string const& indent,
 }
 
 class Interface;
-
-class Argument
-{
-public:
-    Argument(xmlpp::Element const& node)
-        : name{sanitize_name(node.get_attribute_value("name"))},
-          descriptor{parse_optional(node) ? optional_type_map.at(node.get_attribute_value("type"))
-                                          : type_map.at(node.get_attribute_value("type"))}
-    {
-    }
-
-    Emitter emit_c_prototype() const
-    {
-        return {descriptor.c_type, " ", name};
-    }
-
-    Emitter emit_cpp_prototype() const
-    {
-        return {descriptor.cpp_type, " ", name};
-    }
-
-    Emitter emit_thunk_call_fragment() const
-    {
-        return descriptor.converter ? (name + "_resolved") : name;
-    }
-
-    Emitter emit_thunk_converter() const
-    {
-        std::vector<Emitter> substituted_lines;
-        for (auto const& line : descriptor.converter.value_or(std::vector<std::string>{}))
-        {
-            std::string substituted_line = line;
-            size_t substitution_pos = substituted_line.find("$NAME");
-            while (substitution_pos != std::string::npos)
-            {
-                substituted_line = substituted_line.replace(substitution_pos, 5, name);
-                substitution_pos = substituted_line.find("$NAME");
-            }
-            if (!substituted_line.empty())
-                substituted_lines.push_back(substituted_line);
-        }
-        return Lines{substituted_lines};
-    }
-
-private:
-    std::string const name;
-    ArgumentTypeDescriptor const& descriptor;
-};
 
 class Method
 {
